Refactored packages.
Even better package distribution remembering that Activities should be in the main package.
