+++ app/src/main/java/nl/tudelft/pixelperfect/pixelperfect/RoleActivity.java
@@ -7,11 +7,11 @@ import android.view.MenuItem;
 import android.view.View;
 
 import nl.tudelft.pixelperfect.client.GameClient;
-import nl.tudelft.pixelperfect.client.RoleChosenMessage;
-import nl.tudelft.pixelperfect.location.LocationArmoryActivity;
-import nl.tudelft.pixelperfect.location.LocationDeckActivity;
-import nl.tudelft.pixelperfect.location.LocationEngineroomActivity;
-import nl.tudelft.pixelperfect.location.LocationLabActivity;
+import nl.tudelft.pixelperfect.client.message.RoleChosenMessage;
+import nl.tudelft.pixelperfect.pixelperfect.location.LocationArmoryActivity;
+import nl.tudelft.pixelperfect.pixelperfect.location.LocationDeckActivity;
+import nl.tudelft.pixelperfect.pixelperfect.location.LocationEngineroomActivity;
+import nl.tudelft.pixelperfect.pixelperfect.location.LocationLabActivity;
 
 
 /**
