+++ app/src/main/java/nl/tudelft/pixelperfect/pixelperfect/location/LocationArmoryActivity.java
@@ -1,23 +1,24 @@
-package nl.tudelft.pixelperfect.location;
+package nl.tudelft.pixelperfect.pixelperfect.location;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import nl.tudelft.pixelperfect.client.EventCompletedMessage;
+import nl.tudelft.pixelperfect.client.message.EventCompletedMessage;
 import nl.tudelft.pixelperfect.client.GameClient;
 import nl.tudelft.pixelperfect.event.Event;
 import nl.tudelft.pixelperfect.event.Events;
 import nl.tudelft.pixelperfect.pixelperfect.R;
 import nl.tudelft.pixelperfect.pixelperfect.Spaceship;
 
+
 /**
- * This class represents the location of the Engine Room occupied by the Engineer.
+ * This class represents the location of the Armory occupied by the Gunner.
  *
  * @author Floris Doolaard
  */
 @SuppressWarnings({"CanBeFinal", "unused", "UnusedParameters"})
-public class LocationEngineroomActivity extends AppCompatActivity {
+public class LocationArmoryActivity extends AppCompatActivity {
     private GameClient game = GameClient.getInstance();
     private static Spaceship ship = Spaceship.getInstance();
 
@@ -29,7 +30,7 @@ public class LocationEngineroomActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_location_engineroom);
+        setContentView(R.layout.activity_location_armory);
     }
 
     /**
@@ -55,13 +56,13 @@ public class LocationEngineroomActivity extends AppCompatActivity {
     }
 
     /**
-     * Whenever the button to complete the Plasma Leak Event is pressed this will happen.
+     * Whenever the button to complete the Hostile Ship Event is pressed this will happen.
      *
      * @param view , the view of the page.
      */
-    public void completePlasmaLeakEvent(View view){
-        if(ship.getEventLog().contains(Events.PLASMA)){
-            game.sendMessage(new EventCompletedMessage("Plasma Leak Event", ship.getEventLog().pop(Events.PLASMA).getId()));
+    public void completeHostileShipEvent(View view){
+        if(ship.getEventLog().contains(Events.HOSTILE)){
+            game.sendMessage(new EventCompletedMessage("Hostile Ship Event", ship.getEventLog().pop(Events.HOSTILE).getId()));
         } else {
             game.sendMessage(new EventCompletedMessage("WRONG ANSWER", -1));
         }
