+++ app/src/main/java/nl/tudelft/pixelperfect/pixelperfect/location/LocationDeckActivity.java
@@ -1,24 +1,23 @@
-package nl.tudelft.pixelperfect.location;
+package nl.tudelft.pixelperfect.pixelperfect.location;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import nl.tudelft.pixelperfect.client.EventCompletedMessage;
+import nl.tudelft.pixelperfect.client.message.EventCompletedMessage;
 import nl.tudelft.pixelperfect.client.GameClient;
 import nl.tudelft.pixelperfect.event.Event;
 import nl.tudelft.pixelperfect.event.Events;
 import nl.tudelft.pixelperfect.pixelperfect.R;
 import nl.tudelft.pixelperfect.pixelperfect.Spaceship;
 
-
 /**
- * This class represents the location of the Armory occupied by the Gunner.
+ * The location of the Janitor where he can make some coffee.
  *
  * @author Floris Doolaard
  */
-@SuppressWarnings({"CanBeFinal", "unused", "UnusedParameters"})
-public class LocationArmoryActivity extends AppCompatActivity {
+@SuppressWarnings("CanBeFinal")
+public class LocationDeckActivity extends AppCompatActivity {
     private GameClient game = GameClient.getInstance();
     private static Spaceship ship = Spaceship.getInstance();
 
@@ -30,7 +29,7 @@ public class LocationArmoryActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_location_armory);
+        setContentView(R.layout.activity_location_deck);
     }
 
     /**
@@ -50,19 +49,6 @@ public class LocationArmoryActivity extends AppCompatActivity {
     public void completeFireEvent(View view){
         if(ship.getEventLog().contains(Events.FIRE)){
             game.sendMessage(new EventCompletedMessage("Fire Event", ship.getEventLog().pop(Events.FIRE).getId()));
-        } else {
-            game.sendMessage(new EventCompletedMessage("WRONG ANSWER", -1));
-        }
-    }
-
-    /**
-     * Whenever the button to complete the Hostile Ship Event is pressed this will happen.
-     *
-     * @param view , the view of the page.
-     */
-    public void completeHostileShipEvent(View view){
-        if(ship.getEventLog().contains(Events.HOSTILE)){
-            game.sendMessage(new EventCompletedMessage("Hostile Ship Event", ship.getEventLog().pop(Events.HOSTILE).getId()));
         } else {
             game.sendMessage(new EventCompletedMessage("WRONG ANSWER", -1));
         }
