+++ app/src/test/java/nl/tudelft/pixelperfect/pixelperfect/clienttest/RoleChosenMessageTest.java
@@ -4,7 +4,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import nl.tudelft.pixelperfect.client.RoleChosenMessage;
+import nl.tudelft.pixelperfect.client.message.RoleChosenMessage;
 import nl.tudelft.pixelperfect.pixelperfect.Roles;
 
 /**
